Windows: Refactoring for compilation on Windows
