+++ checkpoint_windows.go
@@ -1,0 +1,7 @@
+package main
+
+import (
+	"github.com/codegangsta/cli"
+)
+
+var checkpointCommand = cli.Command{}
