+++ run_windows.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/codegangsta/cli"
+)
+
+func execContainer(context *cli.Context, spec *WindowsSpec) (int, error) {
+	// TODO This will have a different implementation on Windows
+	return 0, nil
+}
