+++ spec_linux.go
@@ -1,5 +1,3 @@
-// +build linux
-
 package main
 
 import (
